Answers to the assignments
Max_Value.java is the code for problem 1: Max Value.
The int[] nums serves as the input array.
The result is generated by running the main() method.
The result is printed on the console.

Grid_Folding.java is the code for problem 2: Grid Folding.
The String 'operation' serves as the input array.
The answer is generated by running the main() method.
The answer is stored in the int[] answer and printed on the console